+++ packages/fe/assets/scss/main.scss
@@ -97,7 +97,7 @@ body {
   padding: 0;
   margin: 0;
   height: auto;
-  color: white;
+  color: black;
   font-size: 16px;
   line-height: 1.2;
   letter-spacing: 0px;
@@ -412,6 +412,15 @@ h6, .h6 {
   }
 }
 
+.no-select {
+  -webkit-touch-callout: none; /* iOS Safari */
+    -webkit-user-select: none; /* Safari */
+     -khtml-user-select: none; /* Konqueror HTML */
+       -moz-user-select: none; /* Old versions of Firefox */
+        -ms-user-select: none; /* Internet Explorer/Edge */
+            user-select: none; /* Non-prefixed version, currently supported by Chrome, Edge, Opera and Firefox */
+}
+
 ///////////////////////////////////////////////////////////////////// Components
 //------------------------------------------------------------------------------
 .grecaptcha-badge {
@@ -490,15 +499,15 @@ h6, .h6 {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  border: 2px solid white;
-  background-color: black;
   padding: 1rem;
   border-radius: 3rem;
+  @include appearanceAnimation;
   .icon {
     display: none;
   }
   .message {
     font-weight: 500;
+    @include linkShadow;
   }
 }
